brightness controls: move screenshot code to its own model
Add a new model to provide a general interface for making
screenshots and refactor globalkey code to use this model.
